Move formatting preferences into appearance tab
Consolidate the Formatting tab contents into the Appearance tab,
placing formatting options after the theme selector and before
the diff section. This simplifies the preferences dialog structure
while keeping all visual/display settings together.

- Remove Formatting from PreferencesTab enum
- Merge formatting controls into Appearance component
- Delete standalone formatting.tsx component
- Update styles to use .formatting-section class
