+++ app/src/ui/preferences/appearance.tsx
@@ -8,14 +8,36 @@ import { Row } from '../lib/row'
 import { DialogContent } from '../dialog'
 import { RadioGroup } from '../lib/radio-group'
 import { Select } from '../lib/select'
+import { Checkbox, CheckboxValue } from '../lib/checkbox'
 import { encodePathAsUrl } from '../../lib/path'
 import { tabSizeDefault } from '../../lib/stores/app-store'
+import { enableFormattingPreferences } from '../../lib/feature-flag'
+import {
+  DateFormat,
+  TimeFormat,
+  INumberFormat,
+  dateFormats,
+  timeFormats,
+  numberFormats,
+  numberFormatToKey,
+} from '../../models/formatting-preferences'
+import { formatNumber } from '../../lib/format-number'
 
 interface IAppearanceProps {
   readonly selectedTheme: ApplicationTheme
   readonly onSelectedThemeChanged: (theme: ApplicationTheme) => void
   readonly selectedTabSize: number
   readonly onSelectedTabSizeChanged: (tabSize: number) => void
+  readonly selectedDateFormat: DateFormat
+  readonly onSelectedDateFormatChanged: (format: DateFormat) => void
+  readonly selectedTimeFormat: TimeFormat
+  readonly onSelectedTimeFormatChanged: (format: TimeFormat) => void
+  readonly selectedNumberFormat: INumberFormat
+  readonly onSelectedNumberFormatChanged: (format: INumberFormat) => void
+  readonly relativeTimeInCommitList: boolean
+  readonly onRelativeTimeInCommitListChanged: (value: boolean) => void
+  readonly relativeTimeInBranchList: boolean
+  readonly onRelativeTimeInBranchListChanged: (value: boolean) => void
 }
 
 interface IAppearanceState {
@@ -76,6 +98,41 @@ export class Appearance extends React.Component<
     event: React.FormEvent<HTMLSelectElement>
   ) => {
     this.props.onSelectedTabSizeChanged(parseInt(event.currentTarget.value))
+  }
+
+  private onDateFormatChanged = (event: React.FormEvent<HTMLSelectElement>) => {
+    this.props.onSelectedDateFormatChanged(
+      event.currentTarget.value as DateFormat
+    )
+  }
+
+  private onTimeFormatChanged = (event: React.FormEvent<HTMLSelectElement>) => {
+    this.props.onSelectedTimeFormatChanged(
+      event.currentTarget.value as TimeFormat
+    )
+  }
+
+  private onNumberFormatChanged = (
+    event: React.FormEvent<HTMLSelectElement>
+  ) => {
+    const match = numberFormats.find(
+      n => numberFormatToKey(n) === event.currentTarget.value
+    )
+    if (match) {
+      this.props.onSelectedNumberFormatChanged(match)
+    }
+  }
+
+  private onRelativeTimeInCommitListChanged = (
+    event: React.FormEvent<HTMLInputElement>
+  ) => {
+    this.props.onRelativeTimeInCommitListChanged(event.currentTarget.checked)
+  }
+
+  private onRelativeTimeInBranchListChanged = (
+    event: React.FormEvent<HTMLInputElement>
+  ) => {
+    this.props.onRelativeTimeInBranchListChanged(event.currentTarget.checked)
   }
 
   public renderThemeSwatch = (theme: ApplicationTheme) => {
@@ -144,13 +201,83 @@ export class Appearance extends React.Component<
     )
   }
 
+  private renderFormatting() {
+    if (!enableFormattingPreferences()) {
+      return null
+    }
+
+    return (
+      <div className="appearance-section formatting-section">
+        <Row>
+          <Select
+            label={__DARWIN__ ? 'Date Format' : 'Date format'}
+            value={this.props.selectedDateFormat}
+            onChange={this.onDateFormatChanged}
+          >
+            {dateFormats.map(({ pattern, example }) => (
+              <option key={pattern} value={pattern}>
+                {example} ({pattern})
+              </option>
+            ))}
+          </Select>
+
+          <Select
+            label={__DARWIN__ ? 'Time Format' : 'Time format'}
+            value={this.props.selectedTimeFormat}
+            onChange={this.onTimeFormatChanged}
+          >
+            {timeFormats.map(({ pattern, example }) => (
+              <option key={pattern} value={pattern}>
+                {example} ({pattern})
+              </option>
+            ))}
+          </Select>
+        </Row>
+
+        <Select
+          label={__DARWIN__ ? 'Number Format' : 'Number format'}
+          value={numberFormatToKey(this.props.selectedNumberFormat)}
+          onChange={this.onNumberFormatChanged}
+        >
+          {numberFormats.map(format => (
+            <option
+              key={numberFormatToKey(format)}
+              value={numberFormatToKey(format)}
+            >
+              {formatNumber(1234567.89, format)}
+            </option>
+          ))}
+        </Select>
+
+        <div className="relative-time-section">
+          <Checkbox
+            label="Show relative time in commit list"
+            value={
+              this.props.relativeTimeInCommitList
+                ? CheckboxValue.On
+                : CheckboxValue.Off
+            }
+            onChange={this.onRelativeTimeInCommitListChanged}
+          />
+          <Checkbox
+            label="Show relative time in branch list"
+            value={
+              this.props.relativeTimeInBranchList
+                ? CheckboxValue.On
+                : CheckboxValue.Off
+            }
+            onChange={this.onRelativeTimeInBranchListChanged}
+          />
+        </div>
+      </div>
+    )
+  }
+
   private renderSelectedTabSize() {
     const availableTabSizes: number[] = [1, 2, 3, 4, 5, 6, 8, 10, 12]
 
     return (
       <div className="appearance-section">
-        <h2 id="diff-heading">{'Diff'}</h2>
-
         <Select
           value={this.state.selectedTabSize.toString()}
           label={__DARWIN__ ? 'Tab Size' : 'Tab size'}
@@ -170,6 +297,7 @@ export class Appearance extends React.Component<
     return (
       <DialogContent>
         {this.renderSelectedTheme()}
+        {this.renderFormatting()}
         {this.renderSelectedTabSize()}
       </DialogContent>
     )
